+++ packages/frontend/package.json
@@ -25,11 +25,11 @@
     "moment": "^2.30.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.1",
-    "react-select": "^5.8.0"
+    "react-router-dom": "^6.22.1"
   },
   "devDependencies": {
     "@esp-group-one/eslint-config": "workspace:^",
+    "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
     "@testing-library/user-event": "^14.5.2",
